Directly run &shell when opening a new terminal
According to the documentation of termopen(), when passing a string to
termopen() as its first argument {cmd}, it spaws &shell and passes it
this string.
In this case, this results in &shell spawning &shell, or bash in bash.

However, if a list is passed to termopen(), it directly executes the
command.
In this case, this results in &shell beings spawned only once.